adding minecard component and interface in type  folder
